replay the current buffer if we're still generating the next
This takes effect if we have played through all the samples in
one buffer before the other has regenerated. It fixes some nasty
distortion when playing too many keys.

It introduces a disconnect in the output when the buffer is
replayed, at about 122Hz given our 15625Hz clock and 128 byte buffers.
It doesn't sound good, but it's better than the previous behavior.